feat:  布局优化、增加短链接更新事件和处理监听器

+++ leopard-server/src/main/java/cn/xilio/leopard/domain/dataobject/ShortUrl.java
@@ -64,7 +64,7 @@ public class ShortUrl extends Auditable {
      * 状态
      */
     @Column(name = "status", nullable = false)
-    private Byte status = 1;
+    private Integer status = 1;
 
     /**
      * 域名
